changed src/devices/timer.c timer_sleep(ticks) changed so that when timer_sleep(ticks) is called, add ticks information(how much to sleep) to the thread, the thread is added to sleep_list, the thread is blocked/and interrupt is disabled while blocking the thread.
changes to be made
the function called by timer_interrupt()
reference: the timer_interrupt function is called when certain no. of
ticks is passed; it is not to be called in the timer_sleep(ticks) function,
calling it will cause interrupt. Therefore, the functions related to
waking the threads should be implemented in the function that is
called by timer_interrupt.
wake up blocked thread,
remove from sleep_list,
add to ready_list,
unblock.